AttentionAgent/requirements.txt
atari-py==0.2.9
box2d-py==2.3.8
cma
gin-config
gym==0.15.3
gym3==0.3.3
matplotlib
opencv-python
procgen==0.10.4
torch
torchvision
#autorom #for Atari Roms: https://pypi.org/project/AutoROM/ and https://github.com/openai/atari-py#roms
